requirements.txt
# requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
cloudinary
django-filter
drf-spectacular # For Swagger/OpenAPI docs
gunicorn # For deployment
psycopg2-binary